requirements.txt
numpy==1.19.5
Flask==2.0.1
scipy==1.5.4
Pillow==8.2.0
Cython==0.29.23
matplotlib==3.3.4
scikit-image==0.17.2
tensorflow==1.15.3
gunicorn
Keras==2.0.8
opencv-contrib-python==4.5.2.52
opencv-python==4.5.2.52
h5py==2.10.0
imgaug==0.4.0
IPython[all]